feat: 支持迁移包含已归档制品的仓库 #3037
1. 修复服务启动失败
