+++ src/backend/job/biz-job/src/main/kotlin/com/tencent/bkrepo/job/migrate/executor/FinishExecutor.kt
@@ -37,7 +37,7 @@ import com.tencent.bkrepo.job.migrate.pojo.MigrateRepoStorageTaskState.FINISHING
 import com.tencent.bkrepo.job.migrate.pojo.MigrateRepoStorageTaskState.MIGRATE_FAILED_NODE_FINISHED
 import com.tencent.bkrepo.job.migrate.pojo.MigrationContext
 import com.tencent.bkrepo.job.migrate.utils.ExecutingTaskRecorder
-import com.tencent.bkrepo.job.service.ArchiveJobService
+import com.tencent.bkrepo.job.service.MigrateArchivedFileService
 import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Component
 
@@ -49,7 +49,7 @@ class FinishExecutor(
     migrateFailedNodeDao: MigrateFailedNodeDao,
     storageService: StorageService,
     executingTaskRecorder: ExecutingTaskRecorder,
-    archiveJobService: ArchiveJobService,
+    migrateArchivedFileService: MigrateArchivedFileService,
     private val repositoryService: RepositoryService,
 ) : BaseTaskExecutor(
     properties,
@@ -58,7 +58,7 @@ class FinishExecutor(
     fileReferenceService,
     storageService,
     executingTaskRecorder,
-    archiveJobService,
+    migrateArchivedFileService,
 ) {
     /**
      * 迁移任务执行结束后对相关资源进行清理
