+++ src/backend/job/biz-job/src/main/kotlin/com/tencent/bkrepo/job/service/ArchiveJobService.kt
@@ -23,16 +23,4 @@ interface ArchiveJobService {
      * 恢复文件，从归档或者压缩中恢复文件
      * */
     fun restore(request: ArchiveRestoreRequest)
-
-    /**
-     * 迁移已归档制品
-     *
-     * @param context 迁移上下文
-     * @param node 待迁移制品
-     *
-     * @return 是否迁移成功，返回true表示迁移成功，false表示文件未归档，抛出异常表示不支持迁移的归档状态
-     *
-     * @throws IllegalStateException 处于不支持迁移的状态时抛出该异常
-     */
-    fun migrateArchivedFile(context: MigrationContext, node: Node): Boolean
 }
