+++ src/backend/job/biz-job/src/main/kotlin/com/tencent/bkrepo/job/service/MigrateArchivedFileService.kt
@@ -25,12 +25,21 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.tencent.bkrepo.archive
+package com.tencent.bkrepo.job.service
 
-import com.tencent.bkrepo.archive.repository.ArchiveFileDao
-import com.tencent.bkrepo.archive.repository.CompressFileDao
-import org.springframework.context.annotation.Import
+import com.tencent.bkrepo.job.migrate.pojo.MigrationContext
+import com.tencent.bkrepo.job.migrate.pojo.Node
 
-
-@Import(ArchiveFileDao::class, CompressFileDao::class)
-class ArchiveAutoConfiguration
+interface MigrateArchivedFileService {
+    /**
+     * 迁移已归档制品
+     *
+     * @param context 迁移上下文
+     * @param node 待迁移制品
+     *
+     * @return 是否迁移成功，返回true表示迁移成功，false表示文件未归档，抛出异常表示不支持迁移的归档状态
+     *
+     * @throws IllegalStateException 处于不支持迁移的状态时抛出该异常
+     */
+    fun migrateArchivedFile(context: MigrationContext, node: Node): Boolean
+}
